Firebase tokens refactor
