+++ lib/providers/services/firebase/firebase_database_map_keys.dart
@@ -1,5 +1,3 @@
-const String FIREBASE_MESSAGE_TOKENS_KEY = 'firebaseMessageTokens';
-
 const String GUARDIANS_COLLECTION_KEY = 'guardians';
 const String GUARDIANS_STATUS_KEY = 'status';
 const String GUARDIANS_DATE_CREATED_KEY = 'dateCreated';
