+++ lib/providers/services/firebase/firebase_database_service.dart
@@ -23,25 +23,6 @@ class FirebaseDatabaseService {
         .map((DocumentSnapshot userData) => FirebaseUser.fromMap(userData.data(), accountName));
   }
 
-  Future<void> setFirebaseMessageToken(String userId) {
-    // Users can have multiple tokens. Ex: Multiple devices.
-    var tokens = <String>[PushNotificationService().token];
-    var data = <String, Object>{
-      FIREBASE_MESSAGE_TOKENS_KEY: FieldValue.arrayUnion(tokens),
-    };
-
-    return _usersCollection.doc(userId).set(data, SetOptions(merge: true));
-  }
-
-  Future<void> removeFirebaseMessageToken(String userId) {
-    var tokens = <String>[PushNotificationService().token];
-    var data = <String, Object>{
-      FIREBASE_MESSAGE_TOKENS_KEY: FieldValue.arrayRemove(tokens),
-    };
-
-    return _usersCollection.doc(userId).set(data, SetOptions(merge: true));
-  }
-
   // Manage guardian Ids
   String _createGuardianId({String currentUserId, String otherUserId}) {
     return currentUserId + '-' + otherUserId;
