+++ lib/v2/screens/import_key/interactor/import_key_bloc.dart
@@ -1,8 +1,8 @@
 import 'package:bloc/bloc.dart';
 import 'package:seeds/providers/notifiers/auth_notifier.dart';
 import 'package:seeds/providers/notifiers/settings_notifier.dart';
-import 'package:seeds/providers/services/firebase/firebase_database_service.dart';
 import 'package:seeds/v2/datasource/local/settings_storage.dart';
+import 'package:seeds/v2/datasource/remote/firebase/firebase_message_token_repository.dart';
 import 'package:seeds/v2/domain-shared/page_state.dart';
 import 'package:seeds/v2/screens/import_key/interactor/mappers/import_key_state_mapper.dart';
 import 'package:seeds/v2/screens/import_key/interactor/usecases/check_private_key_use_case.dart';
@@ -38,7 +38,7 @@ class ImportKeyBloc extends Bloc<ImportKeyEvent, ImportKeyState> {
       settingsStorage.saveAccount(event.account, state.privateKey.toString());
       settingsStorage.privateKeyBackedUp = true;
 
-      await FirebaseDatabaseService().setFirebaseMessageToken(event.account);
+      await FirebaseMessageTokenRepository().setFirebaseMessageToken(event.account);
 
       _authNotifier.resetPasscode();
     }
